+++ lib/src/blocs/page_stack/configuration.dart
@@ -28,4 +28,13 @@ class PageStackConfiguration {
       'pc': PageConfiguration.toJsons(pageConfigurations),
     };
   }
+
+  /// The first non-null page configuration from top.
+  PageConfiguration getTopPageConfiguration() {
+    for (final c in pageConfigurations.reversed) {
+      if (c != null) return c;
+    }
+
+    throw Exception('No page configurations found for this stack.');
+  }
 }