+++ lib/src/models/back_pressed_result_enum.dart
@@ -1,0 +1,4 @@
+enum BackPressedResult {
+  close,
+  keep,
+}