+++ lib/src/route_information_parsers/page_stack.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+import '../blocs/page_stack/configuration.dart';
+
+abstract class PageStackRouteInformationParser extends RouteInformationParser<PageStackConfiguration> {
+  @override
+  RouteInformation? restoreRouteInformation(PageStackConfiguration configuration) {
+    return configuration.getTopPageConfiguration().restoreRouteInformation();
+  }
+}